requirements.txt
click==6.6
Flask==0.11
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pytz==2016.4
PyYAML==3.11
requests==2.10.0
urbanairship==0.7.4
Werkzeug==0.11.10
